Fix issue with fast delete of field content
When a field content is deleted too quickly (before the sync mechanism
    happens), an exception is raised in the Python side.
This behavior doesn't occur if you type a value in a field, wait a
little bit that the sync happens, and delete it.
What's happening is that the code related to the netpyne field tries to
delete a key that doesn't exists yet (see the video below).

The fix removes this behavior (see video below)